NYC 311 Service Requests

Downloads NYC 311 Service Request data from NYC Open Data.

Usage

nyc_311(limit = 10000, filters = list(), timeout_sec = 30)

Source

NYC Open Data: <https://data.cityofnewyork.us/Social-Services/311-Service-Requests-from-2010-to-Present/erm2-nwe9/about_data>

Arguments

limit

Number of rows to retrieve (default = 10,000).

filters

Optional list of field-value pairs to filter results.

timeout_sec

Request timeout in seconds (default = 30).

Value

A tibble containing 311 Service Request data.

Details

This dataset contains all service requests made to NYC 311 since 2010. Data is updated daily and includes agency, complaint type, location, and resolution.

Examples

# Examples that hit the live NYC Open Data API are wrapped so CRAN checks
# do not fail when the network is unavailable or slow.
# \donttest{
if (curl::has_internet()) {
  # Quick example (fetch 2 rows)
  small_sample <- nyc_311(limit = 2)
  small_sample
  nyc_311(filters = list(agency = "NYPD", city = "BROOKLYN"))
}
#> # A tibble: 10,000 × 37
#>    unique_key created_date          agency agency_name complaint_type descriptor
#>    <chr>      <chr>                 <chr>  <chr>       <chr>          <chr>     
#>  1 67573446   2026-01-23T02:19:18.… NYPD   New York C… Noise - Resid… Loud Musi…
#>  2 67572026   2026-01-23T02:18:00.… NYPD   New York C… Illegal Parki… Commercia…
#>  3 67570616   2026-01-23T02:16:37.… NYPD   New York C… Noise - Comme… Loud Musi…
#>  4 67570609   2026-01-23T02:16:26.… NYPD   New York C… Illegal Parki… Blocked H…
#>  5 67574840   2026-01-23T02:11:49.… NYPD   New York C… Noise - Resid… Loud Talk…
#>  6 67572050   2026-01-23T02:10:37.… NYPD   New York C… Noise - Resid… Loud Tele…
#>  7 67580496   2026-01-23T02:09:19.… NYPD   New York C… Noise - Resid… Banging/P…
#>  8 67574821   2026-01-23T02:08:48.… NYPD   New York C… Illegal Parki… Blocked H…
#>  9 67576304   2026-01-23T01:58:32.… NYPD   New York C… Illegal Parki… Parking P…
#> 10 67580406   2026-01-23T01:53:21.… NYPD   New York C… Blocked Drive… No Access 
#> # ℹ 9,990 more rows
#> # ℹ 31 more variables: location_type <chr>, incident_zip <chr>,
#> #   incident_address <chr>, street_name <chr>, cross_street_1 <chr>,
#> #   cross_street_2 <chr>, intersection_street_1 <chr>,
#> #   intersection_street_2 <chr>, address_type <chr>, city <chr>,
#> #   landmark <chr>, status <chr>, community_board <chr>,
#> #   council_district <chr>, police_precinct <chr>, bbl <chr>, borough <chr>, …
# }